Add delete feautre for msme_profile
It will get the value but the var name is profile
then the corresponding sql will be assigned
also for added measurement the one from the msme_management will be
called id_user instead of id

+++ admin/php/msme_delete.php
@@ -2,9 +2,19 @@
 require_once "/opt/lampp/htdocs/busReg_capstone/php/connection.php";
 ini_set('display_errors', 1);
 
-$sql = "DELETE FROM users WHERE id = ?";
+if(isset($_GET['id_user'])){
+    $user_id = urldecode($_GET['id_user']);
+    $sql = "DELETE FROM users WHERE id = ?";
+    $page = "management";
+}else if(isset($_GET['profile'])){
+    $user_id = validate($_GET['profile']);
+    $sql = "DELETE user_profile, new_documents
+        FROM user_profile
+        LEFT JOIN new_documents ON user_profile.user_id = new_documents.user_id
+        WHERE user_profile.user_id = ?";
+    $page = "profile";
+}
 
-$user_id = urldecode($_GET['id']);
 $userDirectory = '/opt/lampp/htdocs/busReg_capstone/user/upload/' . $user_id;
 
 if ($stmt = $mysqli->prepare($sql)) {
@@ -12,8 +22,12 @@ if ($stmt = $mysqli->prepare($sql)) {
     $param_id = $user_id;
     if ($stmt->execute()) {
         echo "User has been deleted";
-        deleteDirectory($userDirectory);
-        header("location: ../msme_management.php");
+        deleteDirectory($userDirectory,$page);
+        if($page === "profile"){
+            header("location: ../msme_profile.php?id=".$user_id);
+        }else if($page === "management") {
+            header("location: ../msme_management.php");
+        }
     } else {
         echo "Error deleting user";
         header("location: ../msme_management.php");
@@ -29,7 +43,7 @@ function validate($data) {
         return $data;
 }
 
-function deleteDirectory($directory) {
+function deleteDirectory($directory,$page) {
     if (!is_dir($directory)) {
         return;
     }
@@ -37,12 +51,16 @@ function deleteDirectory($directory) {
     $files = glob($directory . '/*');
     foreach ($files as $file) {
         if (is_dir($file)) {
-            deleteDirectory($file);
+            deleteDirectory($file,$page);
         } else {
             unlink($file);
         }
     }
 
-    rmdir($directory);
+    if($page === "profile"){
+       
+    }else if($page === "management") {
+        rmdir($directory);
+    }
 }
 ?>
